+++ web-tests/api/src/main/java/ru/instamart/api/model/v1/UserShipmentV1.java
@@ -150,7 +150,7 @@ public class UserShipmentV1 extends BaseObject {
 
     @JsonSchema(required = true)
     @JsonProperty("store_id")
-    private Long storeId;
+    private Integer storeId;
 
     @JsonSchema(required = true)
     private Double total;
